+++ Source/Pawnmorphs/Esoteria/DebugUtils/PMDebugActions.cs
@@ -7,7 +7,9 @@ using System.Linq;
 using System.Text;
 using JetBrains.Annotations;
 using Pawnmorph.Chambers;
+using Pawnmorph.GraphicSys;
 using Pawnmorph.Hediffs;
+using Pawnmorph.Hybrids;
 using Pawnmorph.Jobs;
 using Pawnmorph.Social;
 using Pawnmorph.TfSys;
@@ -17,6 +19,9 @@ using Pawnmorph.Utilities;
 using RimWorld;
 using UnityEngine;
 using Verse;
+using Verse.AI;
+using HarmonyLib;
+using AlienRace;
 
 namespace Pawnmorph.DebugUtils
 {
@@ -41,18 +46,18 @@ namespace Pawnmorph.DebugUtils
 
         }
 
-        [DebugAction(category = PM_CATEGORY,actionType = DebugActionType.Action)]
+        [DebugAction(category = PM_CATEGORY, actionType = DebugActionType.Action)]
         static void GiveBuildupToAllPawns()
         {
             var map = Find.CurrentMap;
-            StringBuilder builder = new StringBuilder(); 
+            StringBuilder builder = new StringBuilder();
             foreach (Pawn pawn in PawnsFinder.AllMaps_SpawnedPawnsInFaction(Faction.OfPlayer).MakeSafe())
             {
-                if(pawn == null) continue;
-                builder.AppendLine(TryGiveMutagenBuildupToPawn(pawn)); 
+                if (pawn == null) continue;
+                builder.AppendLine(TryGiveMutagenBuildupToPawn(pawn));
             }
 
-            Log.Message(builder.ToString()); 
+            Log.Message(builder.ToString());
         }
 
         static string TryGiveMutagenBuildupToPawn(Pawn pawn)
@@ -75,8 +80,8 @@ namespace Pawnmorph.DebugUtils
             StringBuilder sBuilder = new StringBuilder();
             foreach (var kindDef in DefDatabase<PawnKindDef>.AllDefs)
             {
-                var thingDef = kindDef.race; 
-                if(thingDef.race?.Animal != true) continue;
+                var thingDef = kindDef.race;
+                if (thingDef.race?.Animal != true) continue;
 
                 if (!database.TryAddToDatabase(kindDef, out string reason))
                 {
@@ -99,10 +104,10 @@ namespace Pawnmorph.DebugUtils
             if (mutTracker == null)
             {
                 Log.Message("no mutation tracker");
-                return; 
+                return;
             }
 
-            Log.Message(AnimalClassUtilities.GenerateDebugInfo(mutTracker.AllMutations)); 
+            Log.Message(AnimalClassUtilities.GenerateDebugInfo(mutTracker.AllMutations));
 
         }
 
@@ -111,7 +116,7 @@ namespace Pawnmorph.DebugUtils
         {
             var sapienceT = pawn?.GetSapienceTracker();
             if (sapienceT?.CurrentState == null) return;
-            var stateName = sapienceT.CurrentState.StateDef.defName; 
+            var stateName = sapienceT.CurrentState.StateDef.defName;
             try
             {
                 sapienceT.ExitState();
@@ -126,17 +131,17 @@ namespace Pawnmorph.DebugUtils
         static void AddMutation(Pawn pawn)
         {
             if (pawn == null) return;
-            Find.WindowStack.Add(new DebugMenu_AddMutations(pawn)); 
+            Find.WindowStack.Add(new DebugMenu_AddMutations(pawn));
         }
 
 
         [DebugAction(category = PM_CATEGORY, actionType = DebugActionType.Action, allowedGameStates = AllowedGameStates.Playing)]
         static void TagAnimal()
         {
-            
+
             var db = Find.World.GetComponent<ChamberDatabase>();
             var options = GetTaggableAnimalActions(db).ToList();
-            if (options.Count == 0) return; 
+            if (options.Count == 0) return;
             Find.WindowStack.Add(new Dialog_DebugOptionListLister(GetTaggableAnimalActions(db)));
 
         }
@@ -145,14 +150,14 @@ namespace Pawnmorph.DebugUtils
         {
             foreach (PawnKindDef pawnKindDef in DefDatabase<PawnKindDef>.AllDefs)
             {
-                if(!pawnKindDef.race.IsValidAnimal() || db.TaggedAnimals.Contains(pawnKindDef)) continue;
+                if (!pawnKindDef.race.IsValidAnimal() || db.TaggedAnimals.Contains(pawnKindDef)) continue;
                 var tmpPk = pawnKindDef;
-                yield return new DebugMenuOption(pawnKindDef.label,DebugMenuOptionMode.Action,
+                yield return new DebugMenuOption(pawnKindDef.label, DebugMenuOptionMode.Action,
                                                  () => db.AddToDatabase(tmpPk));
-            }   
+            }
         }
 
-        
+
 
 
         static IEnumerable<DebugMenuOption> GetAddMutationOptions([NotNull] Pawn pawn)
@@ -162,31 +167,31 @@ namespace Pawnmorph.DebugUtils
             {
                 if (mDef.parts == null)
                 {
-                    return pawn.health.hediffSet.GetFirstHediffOfDef(mDef) == null; 
+                    return pawn.health.hediffSet.GetFirstHediffOfDef(mDef) == null;
                 }
 
                 foreach (BodyPartDef bodyPartDef in mDef.parts)
                 {
                     foreach (BodyPartRecord record in pawn.GetAllNonMissingParts().Where(p => p.def == bodyPartDef))
                     {
-                        if (!pawn.health.hediffSet.HasHediff(mDef, record)) return true; 
+                        if (!pawn.health.hediffSet.HasHediff(mDef, record)) return true;
                     }
                 }
 
-                return false; 
+                return false;
             }
 
             void CreateMutationDialog(MutationDef mDef)
             {
                 var nMenu = new DebugMenu_AddMutation(mDef, pawn);
-                Find.WindowStack.Add(nMenu); 
+                Find.WindowStack.Add(nMenu);
             }
 
 
             foreach (MutationDef mutation in MutationDef.AllMutations.Where(CanAddMutationToPawn))
             {
                 var tMu = mutation;
-                yield return new DebugMenuOption(mutation.label, DebugMenuOptionMode.Action, () => CreateMutationDialog(tMu)); 
+                yield return new DebugMenuOption(mutation.label, DebugMenuOptionMode.Action, () => CreateMutationDialog(tMu));
             }
 
         }
@@ -206,9 +211,9 @@ namespace Pawnmorph.DebugUtils
         static void ReduceSapience(Pawn pawn)
         {
             var sTracker = pawn?.GetComp<SapienceTracker>();
-            if (sTracker == null) return; 
+            if (sTracker == null) return;
 
-            sTracker.SetSapience(Mathf.Max(0, sTracker.Sapience -0.2f ));
+            sTracker.SetSapience(Mathf.Max(0, sTracker.Sapience - 0.2f));
         }
 
         [DebugAction(category = PM_CATEGORY, actionType = DebugActionType.ToolMapForPawns)]
@@ -228,7 +233,7 @@ namespace Pawnmorph.DebugUtils
             if (pawn.GetSapienceState() != null) return;
             if (!pawn.RaceProps.Animal) return;
 
-            FormerHumanUtilities.MakeAnimalSapient(pawn); 
+            FormerHumanUtilities.MakeAnimalSapient(pawn);
 
         }
         [DebugAction(category = PM_CATEGORY, actionType = DebugActionType.ToolMapForPawns)]
@@ -251,41 +256,369 @@ namespace Pawnmorph.DebugUtils
             TransformedPawn transformedPawn = tfPawn?.pawn;
             if (transformedPawn == null || tfPawn?.status != TransformedStatus.Transformed) return;
             MutagenDef mut = transformedPawn.mutagenDef ?? MutagenDefOf.defaultMutagen;
-            mut.MutagenCached.TryRevert(transformedPawn); 
+            mut.MutagenCached.TryRevert(transformedPawn);
         }
 
 
-        [DebugAction("General","Explosion (mutagenic small)", actionType = DebugActionType.ToolMap, allowedGameStates = AllowedGameStates.PlayingOnMap)]
+        [DebugAction("General", "Explosion (mutagenic small)", actionType = DebugActionType.ToolMap, allowedGameStates = AllowedGameStates.PlayingOnMap)]
         static void SmallExplosionMutagenic()
         {
-            GenExplosion.DoExplosion(UI.MouseCell(), Find.CurrentMap, 10, PMDamageDefOf.MutagenCloud, null); 
+            GenExplosion.DoExplosion(UI.MouseCell(), Find.CurrentMap, 10, PMDamageDefOf.MutagenCloud, null);
         }
 
         [DebugAction("General", "Explosion (mutagenic large)", actionType = DebugActionType.ToolMap, allowedGameStates = AllowedGameStates.PlayingOnMap)]
         static void ExplosionMutagenic()
         {
             GenExplosion.DoExplosion(UI.MouseCell(), Find.CurrentMap, 10, PMDamageDefOf.MutagenCloud_Large, null);
-            
+
         }
 
         [DebugAction("General", "Explosion (mutagenic large dirty)", actionType = DebugActionType.ToolMap, allowedGameStates = AllowedGameStates.PlayingOnMap)]
         static void ExplosionMutagenicDirty()
         {
             GenExplosion.DoExplosion(UI.MouseCell(), Find.CurrentMap, 10, PMDamageDefOf.MutagenCloud_Large, null, postExplosionSpawnThingDef: PMThingDefOf.PM_Filth_Slurry, postExplosionSpawnChance: 0.35f, postExplosionSpawnThingCount: 2);
-            
+
         }
 
-        [DebugAction("Pawnmorpher", "Open action menu", actionType = DebugActionType.Action, allowedGameStates = AllowedGameStates.PlayingOnMap)]
-        public static void OpenActionMenu()
-        {
-            Find.WindowStack.Add(new Pawnmorpher_DebugDialogue());
-        }
-
-        [DebugAction(category=PM_CATEGORY, actionType = DebugActionType.ToolMapForPawns, allowedGameStates = AllowedGameStates.PlayingOnMap)]
+        [DebugAction(category = PM_CATEGORY, actionType = DebugActionType.ToolMapForPawns, allowedGameStates = AllowedGameStates.PlayingOnMap)]
         public static void OpenPartPickerMenu(Pawn pawn)
         {
             if (pawn == null) return;
             Find.WindowStack.Add(new Dialog_PartPicker(pawn, true));
+        }
+
+        private static List<DebugMenuOption> GetRaceChangeOptions()
+        {
+            //var races = RaceGenerator.ImplicitRaces;
+            var lst = new List<DebugMenuOption>();
+            foreach (MorphDef morph in DefDatabase<MorphDef>.AllDefs)
+            {
+                MorphDef local = morph;
+
+                lst.Add(new DebugMenuOption(local.label, DebugMenuOptionMode.Tool, () =>
+                {
+                    Pawn pawn = Find.CurrentMap.thingGrid.ThingsAt(UI.MouseCell()).OfType<Pawn>().FirstOrDefault();
+                    if (pawn != null && pawn.RaceProps.intelligence == Intelligence.Humanlike)
+                        RaceShiftUtilities.ChangePawnToMorph(pawn, local);
+                }));
+            }
+
+            return lst;
+        }
+
+        [DebugAction("Pawnmorpher", "Shift race", actionType = DebugActionType.Action, allowedGameStates = AllowedGameStates.PlayingOnMap)]
+        public static void ShiftRace()
+        {
+            Find.WindowStack.Add(new Dialog_DebugOptionListLister(GetRaceChangeOptions()));
+        }
+
+        private static void GivePawnRandomMutations([CanBeNull] MorphDef morph)
+        {
+            Pawn pawn = Find.CurrentMap.thingGrid
+                            .ThingsAt(UI.MouseCell())
+                            .OfType<Pawn>()
+                            .FirstOrDefault();
+            if (pawn == null) return;
+
+
+            var mutations = morph?.AllAssociatedMutations;
+            if (mutations == null)
+                mutations = DefDatabase<MutationDef>.AllDefs;
+
+
+
+            var mutList = mutations.ToList();
+            if (mutList.Count == 0) return;
+
+            int num = Rand.Range(1, Mathf.Min(10, mutList.Count));
+
+            var i = 0;
+            List<Hediff_AddedMutation> givenList = new List<Hediff_AddedMutation>();
+            List<MutationDef> triedGive = new List<MutationDef>();
+            while (i < num && mutList.Count > 0)
+            {
+                var giver = mutList.RandElement();
+                mutList.Remove(giver);
+                triedGive.Add(giver);
+                var res = MutationUtilities.AddMutation(pawn, giver);
+                givenList.AddRange(res);
+                i++;
+            }
+
+            if (givenList.Count > 0)
+            {
+                Log.Message($"gave {pawn.Name} [{givenList.Join(m => m.Label)}] from [{triedGive.Join(m => m.defName)}]");
+            }
+            else
+            {
+                Log.Message($"could not give {pawn.Name} any from [{triedGive.Join(m => m.defName)}]");
+            }
+        }
+
+        [DebugAction("Pawnmorpher", "Give random mutations", actionType = DebugActionType.Action, allowedGameStates = AllowedGameStates.PlayingOnMap)]
+        private static void GetRandomMutationsOptions()
+        {
+            var options = new List<DebugMenuOption>
+                {new DebugMenuOption("none", DebugMenuOptionMode.Tool, () => GivePawnRandomMutations(null))};
+
+
+            foreach (MorphDef morphDef in MorphDef.AllDefs)
+            {
+                var option = new DebugMenuOption(morphDef.label, DebugMenuOptionMode.Tool,
+                                                 () => GivePawnRandomMutations(morphDef));
+                options.Add(option);
+            }
+
+            Find.WindowStack.Add(new Dialog_DebugOptionListLister(options));
+        }
+
+        [DebugAction("Pawnmorpher", "Recalculate mutation influence", actionType = DebugActionType.ToolMapForPawns, allowedGameStates = AllowedGameStates.PlayingOnMap)]
+        private static void RecalculateInfluence(Pawn pawn)
+        {
+            pawn?.GetComp<MutationTracker>()?.RecalculateMutationInfluences();
+        }
+
+        [DebugAction("Pawnmorpher", "Recalculate all colonist mutation influence", actionType = DebugActionType.Action, allowedGameStates = AllowedGameStates.PlayingOnMap)]
+        private static void AllRecalculateInfluence()
+        {
+            foreach (Pawn colonist in PawnsFinder.AllMaps_FreeColonists)
+            {
+                RecalculateInfluence(colonist);
+            }
+        }
+
+        private static void RaceCheck(Pawn pawn)
+        {
+            if (pawn == null) return;
+            if (pawn.IsAnimalOrMerged()) return;
+            var oldRace = pawn.def;
+            pawn.CheckRace();
+            if (pawn.def == oldRace)
+            {
+                DebugLogUtils.LogMsg(LogLevel.Messages, $"no change in {pawn.Name}");
+            }
+            else
+            {
+                DebugLogUtils.LogMsg(LogLevel.Messages, $"{pawn.Name} was {oldRace.defName} and is now {pawn.def.defName}");
+            }
+        }
+
+        [DebugAction("Pawnmorpher", "Run race check", actionType = DebugActionType.ToolMapForPawns, allowedGameStates = AllowedGameStates.PlayingOnMap)]
+        private static void RunRaceCheck(Pawn pawn)
+        {
+            RaceCheck(pawn);
+        }
+
+        [DebugAction("Pawnmorpher", "Run race check on all pawn", actionType = DebugActionType.Action, allowedGameStates = AllowedGameStates.PlayingOnMap)]
+        private static void RunRaceCheck()
+        {
+            StringBuilder builder = new StringBuilder();
+            foreach (Pawn pawn in PawnsFinder.AllMaps_FreeColonistsAndPrisonersSpawned)
+            {
+                RaceCheck(pawn);
+            }
+        }
+
+        private static void MutationInfo(Pawn pawn)
+        {
+            var tracker = pawn?.GetMutationTracker();
+            if (tracker == null)
+            {
+                Log.Message($"no tracker on {pawn?.Name?.ToStringFull ?? "NULL"}");
+                return;
+            }
+
+            StringBuilder builder = new StringBuilder();
+            builder.AppendLine($"---{pawn.Name}---");
+            builder.AppendLine("---Raw Influence---");
+            foreach (KeyValuePair<AnimalClassBase, float> kvp in tracker)
+            {
+                builder.AppendLine($"{kvp.Key.Label}:{kvp.Value}");
+            }
+
+            builder.AppendLine($"---Total={tracker.TotalInfluence} N:{tracker.TotalNormalizedInfluence} NN:{tracker.TotalInfluence / MorphUtilities.GetMaxInfluenceOfRace(pawn.def)}---");
+
+
+            builder.AppendLine($"---HighestInfluence={tracker.HighestInfluence?.Label ?? "NULL"}---");
+
+
+
+            Log.Message(builder.ToString());
+
+        }
+
+        [DebugAction("Pawnmorpher", "Get mutation info", actionType = DebugActionType.ToolMapForPawns, allowedGameStates = AllowedGameStates.PlayingOnMap)]
+        private static void GetMutationInfo(Pawn pawn)
+        {
+            MutationInfo(pawn);
+        }
+
+
+        [DebugAction("Pawnmorpher", "Get mutation info for all pawns", actionType = DebugActionType.Action, allowedGameStates = AllowedGameStates.PlayingOnMap)]
+        private static void GetAllMutationInfo()
+        {
+            foreach (Pawn pawn in PawnsFinder.AllCaravansAndTravelingTransportPods_Alive)
+            {
+                MutationInfo(pawn);
+            }
+        }
+
+        [DebugAction("Pawnmorpher", "Force full transformation", actionType = DebugActionType.ToolMapForPawns, allowedGameStates = AllowedGameStates.PlayingOnMap)]
+        private static void ForceTransformation(Pawn pawn)
+        {
+            var allHediffs = pawn.health.hediffSet.hediffs;
+            if (allHediffs == null) return;
+
+            foreach (Hediff hediff in allHediffs)
+            {
+                var transformer = hediff.def.GetAllTransformers().FirstOrDefault();
+                if (transformer != null)
+                {
+                    transformer.TransformPawn(pawn, hediff);
+                    return;
+                }
+            }
+        }
+
+        [DebugAction("Pawnmorpher", "Get initial graphics", actionType = DebugActionType.ToolMapForPawns, allowedGameStates = AllowedGameStates.PlayingOnMap)]
+        private static void ListPawnInitialGraphics(Pawn pawn)
+        {
+            var initialComp = pawn.GetComp<InitialGraphicsComp>();
+            if (initialComp == null) return;
+
+            Log.Message(initialComp.GetDebugStr());
+        }
+
+        private static void AddAspectAtStage(AspectDef def, Pawn p, int i)
+        {
+            p.GetAspectTracker()?.Add(def, i);
+        }
+
+        private static void AddBackstoryToPawn(Pawn pawn, BackstoryDef def)
+        {
+            pawn.story.adulthood = def.backstory;
+        }
+
+        private static List<DebugMenuOption> GetGiveBackstoriesOptions(Pawn pawn)
+        {
+            List<DebugMenuOption> options = new List<DebugMenuOption>();
+            foreach (BackstoryDef backstoryDef in DefDatabase<BackstoryDef>.AllDefs)
+            {
+                var def = backstoryDef;
+                options.Add(new DebugMenuOption(def.defName, DebugMenuOptionMode.Action, () => AddBackstoryToPawn(pawn, def)));
+            }
+
+            return options;
+        }
+
+        private static List<DebugMenuOption> GetAddAspectOptions(AspectDef def, Pawn p)
+        {
+            var outLst = new List<DebugMenuOption>();
+            for (var i = 0; i < def.stages.Count; i++)
+            {
+                AspectStage stage = def.stages[i];
+                int i1 = i; //need to make a copy 
+                var label = string.IsNullOrEmpty(stage.label) ? def.label : stage.label;
+                outLst.Add(new DebugMenuOption($"{i}) {label}", DebugMenuOptionMode.Action,
+                                               () => AddAspectAtStage(def, p, i1)));
+            }
+
+            return outLst;
+        }
+
+        private static List<DebugMenuOption> GetAddAspectOptions(Pawn pawn)
+        {
+            AspectTracker tracker = pawn.GetAspectTracker();
+            var outLst = new List<DebugMenuOption>();
+            foreach (AspectDef aspectDef in DefDatabase<AspectDef>.AllDefs.Where(d => !tracker.Contains(d))
+            ) //don't allow aspects to be added more then once 
+            {
+                AspectDef tmpDef = aspectDef;
+
+                outLst.Add(new DebugMenuOption($"{aspectDef.defName}", DebugMenuOptionMode.Action,
+                                               () =>
+                                                   Find.WindowStack
+                                                       .Add(new Dialog_DebugOptionListLister(GetAddAspectOptions(tmpDef,
+                                                                                                                 pawn)))));
+            }
+
+            return outLst;
+        }
+
+        [DebugAction("Pawnmorpher", "Add Aspect", actionType = DebugActionType.ToolMapForPawns, allowedGameStates = AllowedGameStates.PlayingOnMap)]
+        private static void DoAddAspectToPawn(Pawn p)
+        {
+            var options = GetAddAspectOptions(p);
+            if (options.Count == 0) return;
+            Find.WindowStack.Add(new Dialog_DebugOptionListLister(options));
+        }
+
+        private static List<DebugMenuOption> GetRemoveAspectOptions(Pawn p)
+        {
+            var outLst = new List<DebugMenuOption>();
+
+
+            AspectTracker aspectTracker = p.GetAspectTracker();
+            if (aspectTracker == null) return outLst;
+            foreach (Aspect aspect in aspectTracker.Aspects.ToList())
+            {
+                Aspect tmpAspect = aspect;
+                outLst.Add(new DebugMenuOption($"{aspect.Label}", DebugMenuOptionMode.Action,
+                                               () => aspectTracker.Remove(tmpAspect)));
+            }
+
+            return outLst;
+        }
+
+        [DebugAction("Pawnmorpher", "Remove Aspect", actionType = DebugActionType.ToolMapForPawns, allowedGameStates = AllowedGameStates.PlayingOnMap)]
+        private static void DoRemoveAspectsOption(Pawn p)
+        {
+            var options = GetRemoveAspectOptions(p);
+            if (options.Count == 0) return;
+            Find.WindowStack.Add(new Dialog_DebugOptionListLister(options));
+        }
+
+        [DebugAction("Pawnmorpher", "Add Backstory to Sapient Animal", actionType = DebugActionType.ToolMapForPawns, allowedGameStates = AllowedGameStates.PlayingOnMap)]
+        private static void DoAddBackstoryToPawn(Pawn pawn)
+        {
+            if (!pawn.IsFormerHuman()) return;
+
+            Find.WindowStack.Add(new Dialog_DebugOptionListLister(GetGiveBackstoriesOptions(pawn)));
+
+        }
+
+        [DebugAction("Pawnmorpher", "Try Random Hunt", actionType = DebugActionType.ToolMapForPawns, allowedGameStates = AllowedGameStates.PlayingOnMap)]
+        private static void TryStartRandomHunt(Pawn pawn)
+        {
+            if (!pawn.RaceProps.predator) return;
+            var prey = FormerHumanUtilities.FindRandomPreyFor(pawn);
+            if (prey == null) return;
+            var job = new Job(JobDefOf.PredatorHunt, prey)
+            {
+                killIncappedTarget = true
+            };
+
+            pawn.jobs?.StartJob(job, JobCondition.InterruptForced);
+        }
+
+        [DebugAction("Pawnmorpher", "Make pawn permanently feral", actionType = DebugActionType.ToolMapForPawns, allowedGameStates = AllowedGameStates.PlayingOnMap)]
+        private static void MakePawnPermanentlyFeral(Pawn obj)
+        {
+            if (obj?.IsFormerHuman() != true) return;
+
+            FormerHumanUtilities.MakePermanentlyFeral(obj);
+        }
+
+        [DebugAction("Pawnmorpher", "Restart all mutation progression", actionType = DebugActionType.ToolMapForPawns, allowedGameStates = AllowedGameStates.PlayingOnMap)]
+        private static void ResetMutationProgression(Pawn pawn)
+        {
+            var allHediffs = pawn?.health?.hediffSet?.hediffs;
+            if (allHediffs == null) return;
+
+            foreach (Hediff_AddedMutation mutation in allHediffs.OfType<Hediff_AddedMutation>())
+            {
+                mutation.SeverityAdjust?.Restart();
+            }
         }
     }
 }
