+++ Source/Pawnmorphs/Esoteria/Pawnmorph.csproj
@@ -371,7 +371,6 @@
     <Compile Include="DebugUtils\DebugLogUtils.ThoughtListing.cs" />
     <Compile Include="DebugUtils\IDebugString.cs" />
     <Compile Include="DebugUtils\LogLevel.cs" />
-    <Compile Include="DebugUtils\Pawnmorpher_DebugDialogue.cs" />
     <Compile Include="DebugUtils\Pawnmorpher_InteractionWeightLogDialogue.cs" />
     <Compile Include="DefExtensions\AspectGiverExtension.cs" />
     <Compile Include="DefExtensions\AspectRestriction.cs" />
